Split profiler code into modules
* `Profiler.Time`: The sampling CPU time profiler we know and love.

* `Profiler.Memory`: a WIP memory allocation/deallocation profiler
invokable through the exported `@memprofile` macro.  Data is currently
available through the `Profile.Memory.read_and_coalesce_memprofile_data()`
method, which will return `open_chunks`, `closed_chunks` and
`ghost_chunks`, which represent memory that was allocated but not freed,
memory that was allocated and then freed, and memory that was freed but
never allocated, respectively.